% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawsmultitool_functions.R
\name{draws_mean}
\alias{draws_mean}
\title{Calculates mean of get_draws for specified years}
\usage{
draws_mean(draws, pops, locs, year_id)
}
\arguments{
\item{draws}{A get_draws data set.}

\item{pops}{A get_population data set.}

\item{locs}{A get_location_metadata data set.}

\item{year_id}{The specified single year_id.}
}
\value{
A data table with mean for draws, for the specified year.
}
\description{
Calculates mean of get_draws for specified years
}
\examples{
draws_mean(draws = get_draws, pops = get_population, locs = get_location_metadata, year_id = "2015" )
}
