% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawsmultitool_functions.R
\name{draws_percent_change}
\alias{draws_percent_change}
\title{Calculates percent change of get_draws for specified years}
\usage{
draws_percent_change(draws, pops, locs, year_start, year_end)
}
\arguments{
\item{draws}{A get_draws data set.}

\item{pops}{A get_population data set.}

\item{locs}{A get_location_metadata data set.}

\item{year_start}{The desired year_start for percent change.}

\item{year_end}{The desired year_end for percent change.}
}
\value{
A data table with percent change for draws, for the specified years.
}
\description{
Calculates percent change of get_draws for specified years
}
\examples{
draws_percent_change(draws = get_draws, pops = get_population, locs = get_location_metadata, year_start = "1980", year_end = "2015")
}
