% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawsmultitool_functions.R
\name{publish_column_cell}
\alias{publish_column_cell}
\title{Merges Value and Uncertainty values into one single cell}
\usage{
publish_column_cell(col_name, dt, decimals)
}
\arguments{
\item{col_name}{Common column header for all three columns. Format 'col_name_val'.}

\item{dt}{A data table with the three columns you want to merge.}

\item{decimals}{Decimal places.}
}
\value{
A data table with the merged columns specified.
}
\description{
Merges Value and Uncertainty values into one single cell
}
\examples{
publish_column_cell(col_name = "pct", dt = dt, decimals = 0 )
}
